Move chat handlers to model_call module
